Refactor branding manager to keep everything in one directory
To keep everything at one place, moved all branding manager special
things to one module.
