+++ bot/constants.py
@@ -651,35 +651,6 @@ class Event(Enum):
     voice_state_update = "voice_state_update"
 
 
-class Month(IntEnum):
-    JANUARY = 1
-    FEBRUARY = 2
-    MARCH = 3
-    APRIL = 4
-    MAY = 5
-    JUNE = 6
-    JULY = 7
-    AUGUST = 8
-    SEPTEMBER = 9
-    OCTOBER = 10
-    NOVEMBER = 11
-    DECEMBER = 12
-
-    def __str__(self) -> str:
-        return self.name.title()
-
-
-class AssetType(Enum):
-    """
-    Discord media assets.
-
-    The values match exactly the kwarg keys that can be passed to `Guild.edit`.
-    """
-
-    BANNER = "banner"
-    SERVER_ICON = "icon"
-
-
 # Debug mode
 DEBUG_MODE = 'local' in os.environ.get("SITE_URL", "local")
