+++ bot/decorators.py
@@ -1,5 +1,4 @@
 import asyncio
-import functools
 import logging
 import typing as t
 from contextlib import suppress
@@ -9,7 +8,7 @@ from discord import Member, NotFound
 from discord.ext import commands
 from discord.ext.commands import Cog, Context
 
-from bot.constants import Channels, DEBUG_MODE, RedirectOutput
+from bot.constants import Channels, RedirectOutput
 from bot.utils import function
 from bot.utils.checks import in_whitelist_check
 
@@ -153,24 +152,4 @@ def respect_role_hierarchy(member_arg: function.Argument) -> t.Callable:
                 log.trace(f"{func.__name__}: {target.top_role=} < {actor.top_role=}; calling func")
                 await func(*args, **kwargs)
         return wrapper
-    return decorator
-
-
-def mock_in_debug(return_value: t.Any) -> t.Callable:
-    """
-    Short-circuit function execution if in debug mode and return `return_value`.
-
-    The original function name, and the incoming args and kwargs are DEBUG level logged
-    upon each call. This is useful for expensive operations, i.e. media asset uploads
-    that are prone to rate-limits but need to be tested extensively.
-    """
-    def decorator(func: t.Callable) -> t.Callable:
-        @functools.wraps(func)
-        async def wrapped(*args, **kwargs) -> t.Any:
-            """Short-circuit and log if in debug mode."""
-            if DEBUG_MODE:
-                log.debug(f"Function {func.__name__} called with args: {args}, kwargs: {kwargs}")
-                return return_value
-            return await func(*args, **kwargs)
-        return wrapped
     return decorator
