+++ bot/exts/backend/branding/_errors.py
@@ -1,0 +1,2 @@
+class BrandingError(Exception):
+    """Exception raised by the BrandingManager cog."""
